+++ docs/changes.rst
@@ -9,6 +9,8 @@ What's new in 0.8.1
 --------------------
 
 * Fixed issue where app created but not included in enabled_apps causes project start-up failure
+* Top-level application routes now reside in routes.py instead of __init__.py similar to sub-apps.
+* {projname}_initdb renamed to {projname}_populate
 
 What's new in 0.8
 ------------------